+++ apps/daimo-clippy/src/utils/parsers.ts
@@ -37,3 +37,45 @@ export function parseCreateInviteText(text: string): CreateInviteLinkPayload {
 
   return CreateInviteSchema.parse(payload);
 }
+
+export function parseViewInviteStatusText(text: string) {
+  const strippedText = text.trim().split("=");
+
+  if (strippedText[0] !== "link") {
+    throw new Error(`[SLACK-BOT] Unrecognized parameter ${strippedText[0]}`);
+  }
+
+  if (!strippedText[1]) {
+    throw new Error(`[SLACK-BOT] view-invite-status No link provided`);
+  }
+
+  return strippedText[1];
+}
+
+export function parseSetMaxUsesText(text: string) {
+  const strippedText = text.trim();
+  const parts = strippedText.split(" ");
+
+  let url = "";
+  let maxUses: number | undefined = undefined;
+
+  for (const part of parts) {
+    const [key, value] = part.split("=").map((x) => x.trim());
+
+    if (key === "link") {
+      url = value;
+    } else if (key === "max_uses") {
+      maxUses = Number(value);
+    }
+  }
+
+  if (!url) {
+    throw new Error("[SLACK-BOT] /set-max-uses: No invite link specified");
+  }
+
+  if (!maxUses) {
+    throw new Error("[SLACK-BOT] /set-max-uses: No max uses specified");
+  }
+
+  return { url, maxUses };
+}
